// src/types/XykServiceTypes.d.ts
export interface TokenV2Volume {
    /** * The requested chain name eg: `eth-mainnet`. */
    chain_name: string;
    /** * The requested chain ID eg: `1`. */
    chain_id: string;
    /** * The name of the DEX, eg: `uniswap_v2`. */
    dex_name: string;
    /** * Use the relevant `contract_address` to lookup prices, logos, token transfers, etc. */
    contract_address: string;
    /** * The string returned by the `name()` method. */
    contract_name: string;
    total_liquidity: string;
    total_volume_24h: string;
    /** * The contract logo URL. */
    logo_url: string;
    /** * The ticker symbol for this contract. This field is set by a developer and non-unique across a network. */
    contract_ticker_symbol: string;
    /** * Use contract decimals to format the token balance for display purposes - divide the balance by `10^{contract_decimals}`. */
    contract_decimals: number;
    swap_count_24h: number;
    /** * The list of explorers for the token address. */
    explorers: Explorer[];
    /** * The exchange rate for the requested quote currency. */
    quote_rate: number;
    /** * The 24h exchange rate for the requested quote currency. */
    quote_rate_24h: number;
    /** * A prettier version of the exchange rate for rendering purposes. */
    pretty_quote_rate: string;
    /** * A prettier version of the 24h exchange rate for rendering purposes. */
    pretty_quote_rate_24h: string;
    /** * A prettier version of the total liquidity quote for rendering purposes. */
    pretty_total_liquidity_quote: string;
    /** * A prettier version of the 24h volume quote for rendering purposes. */
    pretty_total_volume_24h_quote: string;
    /** * The total liquidity converted to fiat in `quote-currency`. */
    total_liquidity_quote: number;
    /** * The total volume 24h converted to fiat in `quote-currency`. */
    total_volume_24h_quote: number;
    id: any;
    name: any;
    decimals: any;
    symbol: any;
    logo_url: any;
    valueSymbol: any;
    totalFee: any;
    price: any;
    priceC_24h: any;
    totalFeeValue: any;
    fee24: any;
    apy: any;
    feeValue24: any;
    NAVPS: any;
    totalTradeQuantity: any;
    totalTradeValue: any;
    tradeQuantity24: any;
    tradeValue24: any;
    currentQuantity: any;
    currentValue:any;
    isvaluegood: any;
    investQuantity: number;
    investShares: number;
    allNAVPS: number;
    investActualQuantity: number;
    totalInvestValue: number;
    NAVPS: number;
    profit: number;
    APY: number;
    valueSymbol: string;
    earningRate: number;
    isvaluegood: boolean;
    islockgood: boolean;
    address: string;
}